Upstream version 7.2.1~RC51272

# gpg: Signature made Mo 17 Aug 2015 12:14:58 CEST using RSA key ID 251D1DB0
# gpg: Good signature from "Carsten Schoenert <c.schoenert@t-online.de>"
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: B70D FC6F 134F ECFC 011E  62AA 8301 6014 251D 1DB0

+++ common/IStreamAdapter.cpp
@@ -11,14 +11,13 @@
  * license. Therefore any rights, title and interest in our trademarks 
  * remain entirely with us.
  * 
- * Our trademark policy, <http://www.zarafa.com/zarafa-trademark-policy>,
- * allows you to use our trademarks in connection with Propagation and 
- * certain other acts regarding the Program. In any case, if you propagate 
- * an unmodified version of the Program you are allowed to use the term 
- * "Zarafa" to indicate that you distribute the Program. Furthermore you 
- * may use our trademarks where it is necessary to indicate the intended 
- * purpose of a product or service provided you use it in accordance with 
- * honest business practices. For questions please contact Zarafa at 
+ * Our trademark policy (see TRADEMARKS.txt) allows you to use our trademarks
+ * in connection with Propagation and certain other acts regarding the Program.
+ * In any case, if you propagate an unmodified version of the Program you are
+ * allowed to use the term "Zarafa" to indicate that you distribute the Program.
+ * Furthermore you may use our trademarks where it is necessary to indicate the
+ * intended purpose of a product or service provided you use it in accordance
+ * with honest business practices. For questions please contact Zarafa at
  * trademark@zarafa.com.
  *
  * The interactive user interface of the software displays an attribution 
@@ -109,13 +108,15 @@ HRESULT IStreamAdapter::Seek(LARGE_INTEGER dlibMove, DWORD dwOrigin, ULARGE_INTE
 			m_pos = dlibMove.QuadPart;
 	}
 	else if(dwOrigin == SEEK_CUR) {
-		if (dlibMove.QuadPart < 0 && m_pos < -dlibMove.QuadPart)
+		if (dlibMove.QuadPart < 0 &&
+		    m_pos < static_cast<ULONGLONG>(-dlibMove.QuadPart))
 			m_pos = 0;
 		else
 			m_pos += dlibMove.QuadPart;
 	}
 	else if(dwOrigin == SEEK_END) {
-		if (dlibMove.QuadPart < 0 && m_str.size() < -dlibMove.QuadPart)
+		if (dlibMove.QuadPart < 0 &&
+		    m_str.size() < static_cast<ULONGLONG>(-dlibMove.QuadPart))
 			m_pos = 0;
 		else
 			m_pos = m_str.size() + dlibMove.QuadPart;
